+++ cpp/include/tudat-learn/processing/scalers/minmax_scaler.tpp
@@ -34,7 +34,7 @@ void MinMaxScaler<Datum_t, Label_t>::fit(const Dataset<Datum_t, Label_t> &datase
 }
 
 template <typename Datum_t, typename Label_t>
-void MinMaxScaler<Datum_t, Label_t>::fit(const Dataset<Datum_t, Label_t> &dataset, const std::vector<int> &fit_indices) {
+void MinMaxScaler<Datum_t, Label_t>::fit(const Dataset<Datum_t, Label_t> &dataset, const std::vector<size_t> &fit_indices) {
   if(    dataset.size() < 2) throw std::runtime_error("Dataset needs to be of size 2 or greater for the MinMaxScaler use to be possible.");
   if(fit_indices.size() < 2) throw std::runtime_error("Fit_indices needs to be of size 2 or greater for it to be possible to fit MinMaxScaler.");
 
@@ -66,7 +66,7 @@ Dataset<Datum_t, Label_t> MinMaxScaler<Datum_t, Label_t>::transform(Dataset<Datu
 }
 
 template <typename Datum_t, typename Label_t>
-Dataset<Datum_t, Label_t> MinMaxScaler<Datum_t, Label_t>::transform(const Dataset<Datum_t, Label_t> &dataset, const std::vector<int> &transform_indices) const {
+Dataset<Datum_t, Label_t> MinMaxScaler<Datum_t, Label_t>::transform(const Dataset<Datum_t, Label_t> &dataset, const std::vector<size_t> &transform_indices) const {
   Dataset<Datum_t, Label_t> out_dataset;
   out_dataset.reserve(transform_indices.size());
